Work on PID for drive and flywheel

+++ src/PID.cpp
@@ -5,14 +5,14 @@ PID::PID(float nP, float nI, float nD, float nIMAX, float nIMIN) :
         P(nP), I(nI), D(nD), IMAX(nIMAX), IMIN(nIMIN), sum(0), last(0) {
 }
 
-int PID::GetValue(int value, int goal) {
-    int error = goal - value;
-    int pVal = error * P;
+double PID::GetValue(double value, double goal) {
+    double error = goal - value;
+	double pVal = error * P;
     sum += error;
     if (sum > IMAX) sum = IMAX;
     if (sum < IMIN) sum = IMIN;
-    int iVal = I * sum;
-    int dVal = D * (value - last);
+	double iVal = I * sum;
+	double dVal = D * (value - last);
     last = value;
     return pVal + iVal - dVal;
 }
